Reorganizing `resource-access/components`, and removing usage of `@emotion/react`.
