+++ src/modules/getthis/components/GetThisRecord/index.js
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 import getHoldingByBarcode from '../../getHoldingByBarcode';
 import { TrimString } from '../../../core';
 import { RecordFullFormats, FullRecordPlaceholder } from '../../../records';
-import ResourceAccess from '../../../resource-acccess';
+import { ResourceAccess } from '../../../resource-acccess';
 import PropTypes from 'prop-types';
 
 function GetThisHolding ({ record, barcode }) {
